+++ tests/schemas/test_fastavro_paser_schema.py
@@ -128,14 +128,81 @@ def test_one_to_one_repeated_schema():
         latitude: float
         longitude: float
 
-    class Trip(AvroModel):
+        class Meta:
+            namespace = "types.location_type"
 
+    class Trip(AvroModel):
         start_time: datetime.datetime
         start_location: Location
         finish_time: datetime.datetime
         finish_location: Location
+        middle_location: Location
 
     assert parse_schema(Trip.avro_schema_to_python())
+    assert Trip.fake()
+
+
+def test_one_to_one_repeated_schema_in_array():
+    """
+    Test relationship one-to-one with more than once schema
+    """
+
+    class Location(AvroModel):
+        latitude: float
+        longitude: float
+
+        class Meta:
+            namespace = "types.location_type"
+
+    class Trip(AvroModel):
+        start_time: datetime.datetime
+        start_location: Location
+        finish_time: datetime.datetime
+        finish_location: typing.List[Location]
+
+    assert parse_schema(Trip.avro_schema_to_python())
+    assert Trip.fake()
+
+
+def test_one_to_one_repeated_schema_in_map():
+    """
+    Test relationship one-to-one with more than once schema
+    """
+
+    class Location(AvroModel):
+        latitude: float
+        longitude: float
+
+        class Meta:
+            namespace = "types.location_type"
+
+    class Trip(AvroModel):
+        start_time: datetime.datetime
+        start_location: Location
+        finish_time: datetime.datetime
+        finish_location: typing.Dict[str, Location]
+
+    assert parse_schema(Trip.avro_schema_to_python())
+    assert Trip.fake()
+
+
+def test_one_to_many_repeated_schema_in_array_and_map():
+    class User(AvroModel):
+        name: str
+
+        class Meta:
+            schema_doc = False
+            namespace = "types.user"
+
+    class UserAdvance(AvroModel):
+        users: typing.List[User]
+        accounts: typing.Dict[str, User]
+
+        class Meta:
+            schema_doc = False
+
+    assert parse_schema(UserAdvance.avro_schema_to_python())
+    assert UserAdvance.fake()
 
 
 def test_one_to_many_schema():
@@ -271,15 +338,18 @@ def test_schema_array_with_union_types():
     assert parse_schema(ArrayUnionSchema.avro_schema_to_python())
 
 
-def test_schema_with_optional_and_complex_types():
-    class UnionSchema(AvroModel):
-        "Some Unions"
-        first_union: typing.Union[typing.List[str], int]
-        logical_union: typing.Union[typing.List[datetime.datetime], datetime.date, uuid.uuid4]
-        dict_union: typing.Optional[typing.Dict[str, int]]
-        optional_union: typing.Optional[typing.List[str]]
-        dict_union_optional: typing.Optional[typing.Dict[str, int]]
-        optional_union_with_default: typing.Optional[typing.List[str]] = None
-        dict_union_optional_with_default: typing.Optional[typing.Dict[str, int]] = dataclasses.field(
-            default_factory=lambda: [{"test": 1}]
-        )
+def test_namespaces():
+    class C(AvroModel):
+        pass
+
+    class B(AvroModel):
+        c: C
+
+        class Meta:
+            namespace = "my.namespace"
+
+    class A(AvroModel):
+        b1: B
+        b2: B
+
+    parse_schema(A.avro_schema_to_python())